+++ src/libc/ansi/stdio/fread.c
@@ -43,7 +43,7 @@ fread(void *vptr, size_t size, size_t count, FILE *iop)
     while (s > 0) {
       if (iop->_cnt < s) {
 	if (iop->_cnt > 0) {
-	  memcpy(ptr, iop->_ptr, iop->_cnt);
+	  memcpy(ptr, iop->_ptr, (size_t)iop->_cnt);
 	  ptr += iop->_cnt;
 	  s -= iop->_cnt;
 	}
@@ -57,7 +57,7 @@ fread(void *vptr, size_t size, size_t count, FILE *iop)
 	s--;
       }
       if (iop->_cnt >= s) {
-	memcpy(ptr, iop->_ptr, s);
+	memcpy(ptr, iop->_ptr, (size_t)s);
 	iop->_ptr += s;
 	iop->_cnt -= s;
 	return count;
